Reorganize ui files
