+++ Pic2Map/reportDialog.py
@@ -20,7 +20,7 @@ from PyQt6 import QtGui, QtWidgets, QtCore
 from PyQt6.QtGui import *
 from PyQt6.QtCore import *
 from PyQt6.QtWidgets import *
-from .ui_reportGCP import Ui_ReportGCP
+from .ui.ui_reportGCP import Ui_ReportGCP
 from numpy import mean, min, max, std, zeros, sqrt, pi
 # create the dialog for zoom to point
