+++ Pic2Map/ui/ui_d3_virtual.py
@@ -1,0 +1,137 @@
+# -*- coding: utf-8 -*-
+from PyQt6 import QtCore, QtWidgets
+
+
+class Ui_D3(object):
+    def setupUi(self, Ui_D3):
+        Ui_D3.setObjectName("Ui_D3")
+        Ui_D3.resize(763, 511)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(Ui_D3.sizePolicy().hasHeightForWidth())
+        Ui_D3.setSizePolicy(sizePolicy)
+        Ui_D3.setAnimated(False)
+        Ui_D3.setDockOptions(QtWidgets.QMainWindow.AllowTabbedDocks)
+        Ui_D3.setUnifiedTitleAndToolBarOnMac(False)
+        self.centralwidget = QtWidgets.QWidget(Ui_D3)
+        self.centralwidget.setObjectName("centralwidget")
+        Ui_D3.setCentralWidget(self.centralwidget)
+        self.menubar = QtWidgets.QMenuBar(Ui_D3)
+        self.menubar.setGeometry(QtCore.QRect(0, 0, 763, 21))
+        self.menubar.setObjectName("menubar")
+        Ui_D3.setMenuBar(self.menubar)
+        self.statusbar = QtWidgets.QStatusBar(Ui_D3)
+        self.statusbar.setObjectName("statusbar")
+        Ui_D3.setStatusBar(self.statusbar)
+        self.dockWidget = QtWidgets.QDockWidget(Ui_D3)
+        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Fixed, QtWidgets.QSizePolicy.Fixed)
+        sizePolicy.setHorizontalStretch(0)
+        sizePolicy.setVerticalStretch(0)
+        sizePolicy.setHeightForWidth(self.dockWidget.sizePolicy().hasHeightForWidth())
+        self.dockWidget.setSizePolicy(sizePolicy)
+        self.dockWidget.setFeatures(QtWidgets.QDockWidget.NoDockWidgetFeatures)
+        self.dockWidget.setObjectName("dockWidget")
+        self.dockWidgetContents = QtWidgets.QWidget()
+        self.dockWidgetContents.setObjectName("dockWidgetContents")
+        self.verticalLayout = QtWidgets.QVBoxLayout(self.dockWidgetContents)
+        self.verticalLayout.setObjectName("verticalLayout")
+        self.textBrowser = QtWidgets.QTextBrowser(self.dockWidgetContents)
+        self.textBrowser.setObjectName("textBrowser")
+        self.verticalLayout.addWidget(self.textBrowser)
+        self.label = QtWidgets.QLabel(self.dockWidgetContents)
+        self.label.setObjectName("label")
+        self.verticalLayout.addWidget(self.label)
+        self.horizontalSlider = QtWidgets.QSlider(self.dockWidgetContents)
+        self.horizontalSlider.setMinimum(-90)
+        self.horizontalSlider.setMaximum(90)
+        self.horizontalSlider.setOrientation(QtCore.Qt.Orientation.Horizontal)
+        self.horizontalSlider.setTickPosition(QtWidgets.QSlider.TicksAbove)
+        self.horizontalSlider.setObjectName("horizontalSlider")
+        self.verticalLayout.addWidget(self.horizontalSlider)
+        self.label_2 = QtWidgets.QLabel(self.dockWidgetContents)
+        self.label_2.setObjectName("label_2")
+        self.verticalLayout.addWidget(self.label_2)
+        self.horizontalSlider_2 = QtWidgets.QSlider(self.dockWidgetContents)
+        self.horizontalSlider_2.setMinimum(0)
+        self.horizontalSlider_2.setMaximum(180)
+        self.horizontalSlider_2.setOrientation(QtCore.Qt.Orientation.Horizontal)
+        self.horizontalSlider_2.setTickPosition(QtWidgets.QSlider.TicksAbove)
+        self.horizontalSlider_2.setObjectName("horizontalSlider_2")
+        self.verticalLayout.addWidget(self.horizontalSlider_2)
+        self.label_3 = QtWidgets.QLabel(self.dockWidgetContents)
+        self.label_3.setObjectName("label_3")
+        self.verticalLayout.addWidget(self.label_3)
+        self.horizontalSlider_3 = QtWidgets.QSlider(self.dockWidgetContents)
+        self.horizontalSlider_3.setMaximum(100)
+        self.horizontalSlider_3.setOrientation(QtCore.Qt.Orientation.Horizontal)
+        self.horizontalSlider_3.setInvertedControls(False)
+        self.horizontalSlider_3.setTickPosition(QtWidgets.QSlider.NoTicks)
+        self.horizontalSlider_3.setTickInterval(20)
+        self.horizontalSlider_3.setObjectName("horizontalSlider_3")
+        self.verticalLayout.addWidget(self.horizontalSlider_3)
+        self.frame = QtWidgets.QFrame(self.dockWidgetContents)
+        self.frame.setFrameShape(QtWidgets.QFrame.NoFrame)
+        self.frame.setFrameShadow(QtWidgets.QFrame.Plain)
+        self.frame.setLineWidth(0)
+        self.frame.setObjectName("frame")
+        self.horizontalLayout = QtWidgets.QHBoxLayout(self.frame)
+        self.horizontalLayout.setContentsMargins(0, 0, 0, 0)
+        self.horizontalLayout.setSpacing(6)
+        self.horizontalLayout.setObjectName("horizontalLayout")
+        self.saveKMLButton = QtWidgets.QPushButton(self.frame)
+        self.saveKMLButton.setObjectName("saveKMLButton")
+        self.horizontalLayout.addWidget(self.saveKMLButton)
+        self.loadKMLButton = QtWidgets.QPushButton(self.frame)
+        self.loadKMLButton.setObjectName("loadKMLButton")
+        self.horizontalLayout.addWidget(self.loadKMLButton)
+        self.verticalLayout.addWidget(self.frame)
+        self.pushButton = QtWidgets.QPushButton(self.dockWidgetContents)
+        self.pushButton.setObjectName("pushButton")
+        self.verticalLayout.addWidget(self.pushButton)
+        self.dockWidget.setWidget(self.dockWidgetContents)
+        Ui_D3.addDockWidget(QtCore.Qt.DockWidgetArea(2), self.dockWidget)
+
+        self.retranslateUi(Ui_D3)
+        QtCore.QMetaObject.connectSlotsByName(Ui_D3)
+
+    def retranslateUi(self, Ui_D3):
+        _translate = QtCore.QCoreApplication.translate
+        Ui_D3.setWindowTitle(_translate("Ui_D3", "MainWindow"))
+        self.textBrowser.setHtml(_translate("Ui_D3", "<!DOCTYPE HTML PUBLIC \"-//W3C//DTD HTML 4.0//EN\" \"http://www.w3.org/TR/REC-html40/strict.dtd\">\n"
+"<html><head><meta name=\"qrichtext\" content=\"1\" /><style type=\"text/css\">\n"
+"p, li { white-space: pre-wrap; }\n"
+"</style></head><body style=\" font-family:\'MS Shell Dlg 2\'; font-size:8.25pt; font-weight:400; font-style:normal;\">\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt; font-weight:600;\">Position</span></p>\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt;\">Altitude: left button, up and down</span></p>\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt;\">Side: left button, left and right</span></p>\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt;\">Front: Wheel</span></p>\n"
+"<p style=\"-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px; font-size:8pt;\"><br /></p>\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt; font-weight:600;\">Orientation</span></p>\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt;\">Tilt: Right button, up and down</span></p>\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt;\">Heading: Right button, left and right</span></p>\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt;\">Swing: first slider</span></p>\n"
+"<p style=\"-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px; font-size:8pt;\"><br /></p>\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt; font-weight:600;\">Focal</span></p>\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt;\">Field of view: second slider</span></p>\n"
+"<p style=\"-qt-paragraph-type:empty; margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px; font-size:8pt;\"><br /></p>\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt; font-weight:600;\">Transparency</span></p>\n"
+"<p style=\" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;\"><span style=\" font-size:8pt;\">terrain transparency: third slider</span></p></body></html>"))
+        self.label.setText(_translate("Ui_D3", "Swing"))
+        self.label_2.setText(_translate("Ui_D3", "Fiel of view"))
+        self.label_3.setText(_translate("Ui_D3", "Transparecy"))
+        self.saveKMLButton.setText(_translate("Ui_D3", "Save KML Pose"))
+        self.loadKMLButton.setText(_translate("Ui_D3", "Load KML Pose"))
+        self.pushButton.setText(_translate("Ui_D3", "Proceed"))
+
+
+
+
+if __name__ == "__main__":
+    import sys
+    app = QtWidgets.QApplication(sys.argv)
+    Ui_D3 = QtWidgets.QMainWindow()
+    ui = Ui_D3()
+    ui.setupUi(Ui_D3)
+    Ui_D3.show()
+    sys.exit(app.exec_())
